+++ libbpf-rs/CHANGELOG.md
@@ -1,5 +1,7 @@
 Unreleased
 ----------
+- Added `TcHook::get_handle` and `TcHook::get_priority` methods for restoring
+  TcHook object
 - Added `Program::get_fd_by_id` and `Program::get_id_by_fd` methods for restoring
   bpf management data
 - Added `Map::is_pinned` and `Map::get_pin_path` methods for getting map pin status